flask-backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
fsspec==2024.10.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ifcopenshell==0.8.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
kiwisolver==1.4.7
lark==1.2.2
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
mysql-connector-python==9.1.0
networkx==3.4.2
numpy==2.1.3
openai==1.54.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
pyproj==3.7.0
pyroj==0.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
Rtree==1.3.0
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.1
thop==0.1.1.post2209072238
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
trimesh==4.5.2
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.28
ultralytics-thop==2.0.11
urllib3==2.2.3
Werkzeug==3.1.3
